move the function `is_valid_encoding` from init to utils
